pyproject.toml
[project]
name = "flashforge-claude-skill"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "mapbox-earcut>=2.0.0",
    "numpy>=2.0.2",
    "numpy-stl>=3.1.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "shapely>=2.0.7",
    "svgpathtools>=1.7.2",
    "trimesh>=4.11.0",
]
